+++ src/assets/FieldAsset.php
@@ -2,17 +2,7 @@
 
 namespace benf\neo\assets;
 
-use benf\neo\elements\Block;
-use benf\neo\Field;
-use benf\neo\models\BlockTypeGroup;
-use benf\neo\Plugin as Neo;
-use Craft;
-use craft\base\ElementInterface;
-use craft\fieldlayoutelements\CustomField;
-use craft\helpers\Json;
-use craft\models\FieldLayout;
 use craft\web\AssetBundle;
-use craft\web\assets\cp\CpAsset;
 
 /**
  * Class FieldAsset
@@ -20,7 +10,9 @@ use craft\web\assets\cp\CpAsset;
  * @package benf\neo\assets
  * @author Spicy Web <plugins@spicyweb.com.au>
  * @since 2.0.0
+ * @deprecated in 3.0.0; users of `EVENT_FILTER_BLOCK_TYPES` should use `InputAsset` instead
  */
 class FieldAsset extends AssetBundle
 {
+    public const EVENT_FILTER_BLOCK_TYPES = 'filterBlockTypes';
 }