+++ CHANGELOG.md
@@ -8,6 +8,9 @@
 - Whether a Neo block type's max blocks setting has been exceeded will now be validated server-side when saving a Neo field's contents, rather than relying on it to be enforced by client-side JavaScript
 - Updated JavaScript dependencies
 
+### Deprecated
+- Deprecated `benf\neo\assets\FieldAsset`; users of `EVENT_FILTER_BLOCK_TYPES` should use `InputAsset` instead
+
 ### Removed
 - Removed the unused `saveModifiedBlocksOnly` setting
 - Removed `benf\neo\Field::$localizeBlocks`; use `$propagationMethod` instead